Update hyperkit go module for disk changes
The main benefit is that this will allow us to support disk resizing on
macOS. This will need a bit of extra work in crc as disk resizing requires
the qcow-tool binary to be available.
Now that the driver code doesn't rely on hyperkit.json to find the
hyperkit pid, only 2 minor changes are needed for this update to account
for API changes in hyperkit go bindings.