Separate 'VM is running' check from 'get IP' loop
At the moment, right after starting hyperkit, we have a `getIP` loop
which checks that the VM is running, and then tries to get the IP.
This commit reworks a bit this check to first have a loop checking the
VM is running, and then have a loop getting the VM IP.

The reason for that is that in the following commits we'll change how
we get the hyperkit pid. Currently, it's read from a hyperkit.json file
which is created synchronously hyperkit.Start(). After the following
commits, it will be read from the hyperkit.pid file, which is created
by the hyperkit process, which runs in the background.
This means that after this change is done, we are no longer guaranteed
to be able to get the hyperkit pid right after hyperkit.Start() returns,
so this short loop for GetState() == state.Running is needed to avoid
some startup races.